// WADNR.Web/src/app/shared/generated/model/interaction-event-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { InteractionEventTypeLookupItem } from './interaction-event-type-lookup-item';
import { PersonLookupItem } from './person-lookup-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class InteractionEventGridRow { 
    InteractionEventID?: number;
    InteractionEventTitle?: string | null;
    InteractionEventDescription?: string | null;
    InteractionEventDate?: string;
    StaffPerson?: PersonLookupItem;
    InteractionEventType?: InteractionEventTypeLookupItem;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface InteractionEventGridRowForm { 
    InteractionEventID?: FormControl<number>;
    InteractionEventTitle?: FormControl<string>;
    InteractionEventDescription?: FormControl<string>;
    InteractionEventDate?: FormControl<string>;
    StaffPerson?: FormControl<PersonLookupItem>;
    InteractionEventType?: FormControl<InteractionEventTypeLookupItem>;
}

export class InteractionEventGridRowFormControls { 
    public static InteractionEventID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InteractionEventTitle = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InteractionEventDescription = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InteractionEventDate = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static StaffPerson = (value: FormControlState<PersonLookupItem> | PersonLookupItem = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<PersonLookupItem>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static InteractionEventType = (value: FormControlState<InteractionEventTypeLookupItem> | InteractionEventTypeLookupItem = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<InteractionEventTypeLookupItem>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}